requirements.txt
django==2.1.7
djangorestframework==3.9.2
graphene-django==2.3.2
django-filter==2.1.0
whitenoise==4.1.2
gunicorn==19.9.0
Pillow==5.0.0
environs==4.1.0
django-cors-headers==3.0.2
djangorestframework-simplejwt-4.3.0